feat: refactor everything
