+++ src/core/mod.rs
@@ -15,6 +15,5 @@
  *    limitations under the License.
  */
 
-/// Utility features used within the fleet application
-pub mod bloat;
-pub mod configure;
+pub mod commands;
+pub mod config;
