+++ src/main.rs
@@ -15,25 +15,36 @@
  *    limitations under the License.
  */
 
-//! 
+//!
 //! Fleet is the blazing fast build tool for Rust. Compiling with Fleet is up-to 5x faster than with cargo.
 //!
 //! Note: Since fleet is in the beta phase, it might not be completely stable yet. Feel free to open any issues or bug reports at issues.
 //!
 //! Note: As of now fleet only supports rustc nightly
-//! 
+//!
 #![warn(clippy::all)]
 #![warn(clippy::pedantic)]
-#![allow(clippy::missing_errors_doc, clippy::module_name_repetitions)]
+#![allow(
+    clippy::missing_errors_doc,
+    clippy::module_name_repetitions,
+    clippy::unused_self
+)]
 
-use cli::CLI;
+use anyhow::Result;
+use cli::app::App;
+use human_panic::setup_panic;
 
 pub mod cli;
-pub mod commands;
-pub mod config;
+pub mod core;
 pub mod utils;
 
 /// Entrypoint to the CLI application
-fn main() {
-    CLI::run();
+fn main() -> Result<()> {
+    // Human panic messages are only shown in `release` mode
+    setup_panic!();
+
+    #[cfg(windows)]
+    let _ = ansi_term::enable_ansi_support();
+
+    App::new().run()
 }
