+++ src/utils/configure.rs
@@ -1,13 +1,29 @@
+/*
+ *
+ *    Copyright 2021 Fleet Contributors
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
 use std::env::consts::OS;
 
 use colored::Colorize;
 
-
 /// Installs a linker
-/// 
-/// Installs the specified linker using the appropriate 
+///
+/// Installs the specified linker using the appropriate
 /// package manager for the user operating system
-/// 
+///
 /// # Panics
 /// Can panic if unable to install clang and lld
 pub fn install_linker(linker: &str) {
